Move pncUpdateDataset schema into root schemas folder
