+++ packages/core/conductor-tasks/bichard_phase_2/processPhase2.ts
@@ -15,8 +15,8 @@ import type { PncUpdateDataset } from "../../types/PncUpdateDataset"
 
 import CoreAuditLogger from "../../lib/CoreAuditLogger"
 import phase2 from "../../phase2/phase2"
-import pncUpdateDatasetSchema from "../../phase2/schemas/pncUpdateDataset"
 import { Phase2ResultType } from "../../phase2/types/Phase2Result"
+import pncUpdateDatasetSchema from "../../schemas/pncUpdateDataset"
 import { unvalidatedHearingOutcomeSchema } from "../../schemas/unvalidatedHearingOutcome"
 import { isPncUpdateDataset } from "../../types/PncUpdateDataset"
