+++ packages/core/conductor-tasks/bichard_phase_3/processPhase3.ts
@@ -15,8 +15,8 @@ import type { PncUpdateDataset } from "../../types/PncUpdateDataset"
 import CoreAuditLogger from "../../lib/CoreAuditLogger"
 import createPncApiConfig from "../../lib/createPncApiConfig"
 import PncGateway from "../../lib/PncGateway"
-import pncUpdateDatasetSchema from "../../phase2/schemas/pncUpdateDataset"
 import phase3 from "../../phase3/phase3"
+import pncUpdateDatasetSchema from "../../schemas/pncUpdateDataset"
 
 const pncApiConfig = createPncApiConfig()
 const s3Config = createS3Config()
