+++ packages/core/phase2/schemas/annotatedPncUpdateDataset.ts
@@ -1,6 +1,6 @@
 import { z } from "zod"
 
-import pncUpdateDatasetSchema from "./pncUpdateDataset"
+import pncUpdateDatasetSchema from "../../schemas/pncUpdateDataset"
 
 export const annotatedPNCUpdateDatasetSchema = z.object({
   AnnotatedPNCUpdateDataset: z.object({
