+++ packages/core/phase2/schemas/phase2Result.ts
@@ -1,9 +1,9 @@
 import { auditLogEventSchema } from "@moj-bichard7/common/schemas/auditLogEvent"
 import { z } from "zod"
 
+import pncUpdateDatasetSchema from "../../schemas/pncUpdateDataset"
 import { triggerSchema } from "../../schemas/trigger"
 import { Phase2ResultType } from "../types/Phase2Result"
-import pncUpdateDatasetSchema from "./pncUpdateDataset"
 
 export const phase2ResultSchema = z.object({
   auditLogEvents: z.array(auditLogEventSchema),
