+++ packages/core/phase2/tests/helpers/PartialPncUpdateDataset.ts
@@ -1,6 +1,6 @@
 import type { z } from "zod"
 
-import pncUpdateDatasetSchema from "../../schemas/pncUpdateDataset"
+import pncUpdateDatasetSchema from "../../../schemas/pncUpdateDataset"
 
 const _partialPncUpdateDataset = pncUpdateDatasetSchema.deepPartial()
 export type PartialPncUpdateDataset = z.infer<typeof _partialPncUpdateDataset>
