+++ packages/core/phase3/schemas/phase3Result.ts
@@ -1,7 +1,7 @@
 import { auditLogEventSchema } from "@moj-bichard7/common/schemas/auditLogEvent"
 import { z } from "zod"
 
-import pncUpdateDatasetSchema from "../../phase2/schemas/pncUpdateDataset"
+import pncUpdateDatasetSchema from "../../schemas/pncUpdateDataset"
 import { triggerSchema } from "../../schemas/trigger"
 import { Phase3ResultType } from "../types/Phase3Result"
